Put DatabaseMemory class into separated file.
