+++ dbms/include/DB/Databases/DatabaseMemory.h
@@ -1,29 +1,45 @@
 #pragma once
 
-#include <DB/Databases/DatabaseMemory.h>
+#include <DB/Databases/IDatabase.h>
 
 
 namespace DB
 {
 
-/** Default engine of databases.
-  * It stores tables list in filesystem using list of .sql files,
-  *  that contain declaration of table represented by SQL ATTACH TABLE query.
+/** A non-persistent database to store temporary data.
+  * It doesn't make any manipulations with filesystem.
+  * All tables are created by calling code.
+  * TODO: Maybe DatabaseRuntime is more suitable class name.
   */
-class DatabaseOrdinary : public DatabaseMemory
+class DatabaseMemory : public IDatabase
 {
 protected:
-	const String path;
+	const String name;
+	mutable std::mutex mutex;
+	Tables tables;
+
+	Logger * log;
 
 public:
-	DatabaseOrdinary(const String & name_, const String & path_);
 
-	String getEngineName() const override { return "Ordinary"; }
+	DatabaseMemory(const String & name_) : name(name_) {}
+
+	String getEngineName() const override { return "Memory"; }
 
 	void loadTables(Context & context, ThreadPool * thread_pool, bool has_force_restore_data_flag) override;
 
+	bool isTableExist(const String & table_name) const override;
+	StoragePtr tryGetTable(const String & table_name) override;
+
+	DatabaseIteratorPtr getIterator() override;
+
+	bool empty() const override;
+
 	void createTable(const String & table_name, const StoragePtr & table, const ASTPtr & query, const String & engine) override;
 	void removeTable(const String & table_name) override;
+
+	void attachTable(const String & table_name, const StoragePtr & table) override;
+	StoragePtr detachTable(const String & table_name) override;
 
 	void renameTable(const Context & context, const String & table_name, IDatabase & to_database, const String & to_table_name) override;
