+++ dbms/include/DB/Databases/DatabasesCommon.h
@@ -3,7 +3,7 @@
 #include <DB/Core/Types.h>
 #include <DB/Parsers/IAST.h>
 #include <DB/Storages/IStorage.h>
-
+#include <DB/Databases/IDatabase.h>
 
 /// Общая функциональность для нескольких разных движков баз данных.
 
@@ -30,5 +30,37 @@ std::pair<String, StoragePtr> createTableFromDefinition(
 	bool has_force_restore_data_flag,
 	const String & description_for_error_message);
 
+
+/// Copies list of tables and iterates through such snapshot.
+class DatabaseSnaphotIterator : public IDatabaseIterator
+{
+private:
+	Tables tables;
+	Tables::iterator it;
+
+public:
+	DatabaseSnaphotIterator(Tables & tables_)
+		: tables(tables_), it(tables.begin()) {}
+
+	void next() override
+	{
+		++it;
+	}
+
+	bool isValid() const override
+	{
+		return it != tables.end();
+	}
+
+	const String & name() const override
+	{
+		return it->first;
+	}
+
+	StoragePtr & table() const override
+	{
+		return it->second;
+	}
+};
 
 }
