+++ dbms/src/Databases/DatabaseOrdinary.cpp
@@ -2,6 +2,7 @@
 #include <Poco/DirectoryIterator.h>
 
 #include <DB/Databases/DatabaseOrdinary.h>
+#include <DB/Databases/DatabaseMemory.h>
 #include <DB/Databases/DatabasesCommon.h>
 #include <DB/Common/escapeForFileName.h>
 #include <DB/Common/StringUtils.h>
@@ -89,7 +90,7 @@ static void loadTable(
 
 DatabaseOrdinary::DatabaseOrdinary(
 	const String & name_, const String & path_)
-	: name(name_), path(path_)
+	: DatabaseMemory(name_), path(path_)
 {
 }
 
@@ -187,79 +188,6 @@ void DatabaseOrdinary::loadTables(Context & context, ThreadPool * thread_pool, b
 }
 
 
-bool DatabaseOrdinary::isTableExist(const String & table_name) const
-{
-	std::lock_guard<std::mutex> lock(mutex);
-	return tables.count(table_name);
-}
-
-
-StoragePtr DatabaseOrdinary::tryGetTable(const String & table_name)
-{
-	std::lock_guard<std::mutex> lock(mutex);
-	auto it = tables.find(table_name);
-	if (it == tables.end())
-		return {};
-	return it->second;
-}
-
-
-/// Копирует список таблиц. Таким образом, итерируется по их снапшоту.
-class DatabaseOrdinaryIterator : public IDatabaseIterator
-{
-private:
-	Tables tables;
-	Tables::iterator it;
-
-public:
-	DatabaseOrdinaryIterator(Tables & tables_)
-		: tables(tables_), it(tables.begin()) {}
-
-	void next() override
-	{
-		++it;
-	}
-
-	bool isValid() const override
-	{
-		return it != tables.end();
-	}
-
-	const String & name() const override
-	{
-		return it->first;
-	}
-
-	StoragePtr & table() const override
-	{
-		return it->second;
-	}
-};
-
-
-DatabaseIteratorPtr DatabaseOrdinary::getIterator()
-{
-	std::lock_guard<std::mutex> lock(mutex);
-	return std::make_unique<DatabaseOrdinaryIterator>(tables);
-}
-
-
-bool DatabaseOrdinary::empty() const
-{
-	std::lock_guard<std::mutex> lock(mutex);
-	return tables.empty();
-}
-
-
-void DatabaseOrdinary::attachTable(const String & table_name, const StoragePtr & table)
-{
-	/// Добавляем таблицу в набор.
-	std::lock_guard<std::mutex> lock(mutex);
-	if (!tables.emplace(table_name, table).second)
-		throw Exception("Table " + name + "." + table_name + " already exists.", ErrorCodes::TABLE_ALREADY_EXISTS);
-}
-
-
 void DatabaseOrdinary::createTable(const String & table_name, const StoragePtr & table, const ASTPtr & query, const String & engine)
 {
 	/// Создаём файл с метаданными, если нужно - если запрос не ATTACH.
@@ -313,23 +241,6 @@ void DatabaseOrdinary::createTable(const String & table_name, const StoragePtr &
 		Poco::File(table_metadata_tmp_path).remove();
 		throw;
 	}
-}
-
-
-StoragePtr DatabaseOrdinary::detachTable(const String & table_name)
-{
-	StoragePtr res;
-
-	{
-		std::lock_guard<std::mutex> lock(mutex);
-		auto it = tables.find(table_name);
-		if (it == tables.end())
-			throw Exception("Table " + name + "." + table_name + " doesn't exist.", ErrorCodes::TABLE_ALREADY_EXISTS);
-		res = it->second;
-		tables.erase(it);
-	}
-
-	return res;
 }
 
 
@@ -507,113 +418,5 @@ void DatabaseOrdinary::alterTable(
 		throw;
 	}
 }
-
-bool DatabaseMemory::isTableExist(const String & table_name) const
-{
-	std::lock_guard<std::mutex> lock(mutex);
-	return tables.count(table_name);
-}
-
-StoragePtr DatabaseMemory::tryGetTable(const String & table_name)
-{
-	std::lock_guard<std::mutex> lock(mutex);
-	auto it = tables.find(table_name);
-	if (it == tables.end())
-		return {};
-	return it->second;
-}
-
-DatabaseIteratorPtr DatabaseMemory::getIterator()
-{
-	std::lock_guard<std::mutex> lock(mutex);
-	return std::make_unique<DatabaseOrdinaryIterator>(tables);
-}
-
-bool DatabaseMemory::empty() const
-{
-	std::lock_guard<std::mutex> lock(mutex);
-	return tables.empty();
-}
-
-void DatabaseMemory::createTable(const String & table_name, const StoragePtr & table, const ASTPtr & query, const String & engine)
-{
-	LOG_DEBUG(log, "Creating Table " << table_name);
-	attachTable(table_name, table);
-	LOG_DEBUG(log, "Created Table " << table_name);
-}
-
-void DatabaseMemory::removeTable(const String & table_name)
-{
-	detachTable(table_name);
-}
-
-StoragePtr DatabaseMemory::detachTable(const String & table_name)
-{
-	StoragePtr res;
-	{
-		std::lock_guard<std::mutex> lock(mutex);
-		auto it = tables.find(table_name);
-		if (it == tables.end())
-			throw Exception("Table " + name + "." + table_name + " doesn't exist.", ErrorCodes::TABLE_ALREADY_EXISTS);
-		res = it->second;
-		tables.erase(it);
-	}
-
-	return res;
-}
-
-void DatabaseMemory::attachTable(const String & table_name, const StoragePtr & table)
-{
-	std::lock_guard<std::mutex> lock(mutex);
-	if (!tables.emplace(table_name, table).second)
-		throw Exception("Table " + name + "." + table_name + " already exists.", ErrorCodes::TABLE_ALREADY_EXISTS);
-}
-
-void DatabaseMemory::renameTable(
-	const Context & context, const String & table_name, IDatabase & to_database, const String & to_table_name)
-{
-	throw Exception("renameTable() is not supported", ErrorCodes::NOT_IMPLEMENTED);
-}
-
-time_t DatabaseMemory::getTableMetadataModificationTime(const String & table_name)
-{
-	return static_cast<time_t>(0);
-}
-
-ASTPtr DatabaseMemory::getCreateQuery(const String & table_name) const
-{
-	throw Exception("getCreateQuery() is not supported", ErrorCodes::NOT_IMPLEMENTED);
-	return nullptr;
-}
-
-void DatabaseMemory::shutdown()
-{
-	/// Нельзя удерживать блокировку во время shutdown.
-	/// Потому что таблицы могут внутри функции shutdown работать с БД, а mutex не рекурсивный.
-
-	for (auto iterator = getIterator(); iterator->isValid(); iterator->next())
-		iterator->table()->shutdown();
-
-	std::lock_guard<std::mutex> lock(mutex);
-	tables.clear();
-}
-
-void DatabaseMemory::drop()
-{
-	/// Дополнительных действий по удалению не требуется.
-}
-
-void DatabaseMemory::alterTable(
-	const Context & context,
-	const String & name,
-	const NamesAndTypesList & columns,
-	const NamesAndTypesList & materialized_columns,
-	const NamesAndTypesList & alias_columns,
-	const ColumnDefaults & column_defaults,
-	const ASTModifier & engine_modifier)
-{
-	throw Exception("alterTable() is not supported", ErrorCodes::NOT_IMPLEMENTED);
-}
-
 
 }
