+++ dbms/src/Server/Server.cpp
@@ -519,6 +519,3 @@ int Server::main(const std::vector<std::string> & args)
 }
 
 }
-
-
-YANDEX_APP_SERVER_MAIN(DB::Server);
